\name{KLMMAL-package}
\alias{KLMMAL-package}
\alias{KLMMAL}
\docType{package}
\title{
\packageTitle{KLMMAL}
}
\description{
\packageDescription{KLMMAL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KLMMAL}
\packageIndices{KLMMAL}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{KLMMAL}

Maintainer: \packageMaintainer{KLMMAL}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
